Decompile Prologue: Space Dance.

+++ src/scenes/karate_man.inc.c
@@ -5,7 +5,7 @@
 extern s16 D_03004afc;
 
 // !TODO
-extern void func_0804d770(u32, u32, u16);
+extern void func_0804d770(s32, s16, u16);
 extern void func_0804cebc(s32, s16, s8);
 extern u32  func_0804d160(s32, u32 *, s8, s16, s16, u16, s8, s8, u16);
